♻️ refactor: split large operation files into focused, maintainable modules
Address architectural reviewer's concerns about file size and maintainability
while preserving all functionality and backward compatibility:

Stock Operations Refactoring:
- stock-operations.js: 695+ lines → 108 lines (84% reduction)
- Split into 4 domain-focused modules:
  * stock-inventory.js (263 lines): Stock levels, entries, volatile stock
  * stock-consumption.js (192 lines): Consuming products, opening items
  * stock-purchases.js (140 lines): Adding stock, purchase operations
  * stock-transfers.js (100 lines): Location transfers

Generic Operations Refactoring:
- generic-operations.js: 343+ lines → 101 lines (71% reduction)
- Split into 4 logical modules:
  * entity-operations.js (258 lines): CRUD operations, users, recipes
  * file-operations.js (59 lines): File uploads, downloads, deletions
  * system-operations.js (58 lines): System info, configuration
  * settings-operations.js (57 lines): User settings management

Architecture Benefits:
- Enhanced maintainability with logical domain grouping
- Easier debugging by isolating functionality to specific modules
- Future extensibility with clear patterns for new operations
- Preserved coordination pattern: main files act as module coordinators
- Zero breaking changes: all existing operation names continue working

All 56 operations verified functional with improved code organization.
