+++ nodes/lib/operations/generic-operations.js
@@ -1,11 +1,54 @@
-const Validators = require('../validators');
+const EntityOperations = require('./entity-operations');
+const FileOperations = require('./file-operations');
+const SystemOperations = require('./system-operations');
+const SettingsOperations = require('./settings-operations');
 
 /**
- * Generic entity and system operations
+ * Generic entity and system operations - now coordinates between specialized modules
  */
 class GenericOperations {
   constructor(client) {
+    this.client = client;
     this.api = client.getAPI();
+    
+    // Initialize specialized operation modules
+    this.entities = new EntityOperations(client);
+    this.files = new FileOperations(client);
+    this.system = new SystemOperations(client);
+    this.settings = new SettingsOperations(client);
+    
+    // Build operation mapping to modules
+    this.operationModuleMap = this.buildOperationModuleMap();
+  }
+
+  /**
+   * Build a mapping of operations to their specialized modules
+   * @returns {Map} Operation to module mapping
+   */
+  buildOperationModuleMap() {
+    const map = new Map();
+    
+    // Entity operations
+    this.entities.getSupportedOperations().forEach(op => {
+      map.set(op, this.entities);
+    });
+    
+    // File operations
+    this.files.getSupportedOperations().forEach(op => {
+      map.set(op, this.files);
+    });
+    
+    // System operations
+    this.system.getSupportedOperations().forEach(op => {
+      map.set(op, this.system);
+    });
+    
+    // Settings operations
+    this.settings.getSupportedOperations().forEach(op => {
+      map.set(op, this.settings);
+    });
+    
+    return map;
   }
 
   /**
@@ -17,289 +60,26 @@ class GenericOperations {
    * @returns {Promise<*>} Operation result
    */
   async execute(operation, payload, options = {}, entityType = null) {
-    switch (operation) {
-      // System operations
-      case 'getSystemInfo':
-        return await this.getSystemInfo();
-      
-      case 'getDbChangedTime':
-        return await this.getDbChangedTime();
-      
-      case 'getConfig':
-        return await this.getConfig();
-      
-      case 'getTime':
-        return await this.getTime(payload);
-      
-      // Generic entity operations
-      case 'getObjects':
-        return await this.getObjects(payload, options, entityType);
-      
-      case 'addObject':
-        return await this.addObject(payload, entityType);
-      
-      case 'getObject':
-        return await this.getObject(payload, entityType);
-      
-      case 'editObject':
-        return await this.editObject(payload, entityType);
-      
-      case 'deleteObject':
-        return await this.deleteObject(payload, entityType);
-      
-      // User fields operations
-      case 'getUserfields':
-        return await this.getUserfields(payload);
-      
-      case 'setUserfields':
-        return await this.setUserfields(payload);
-      
-      // File operations
-      case 'getFile':
-        return await this.getFile(payload);
-      
-      case 'uploadFile':
-        return await this.uploadFile(payload);
-      
-      case 'deleteFile':
-        return await this.deleteFile(payload);
-      
-      // User operations
-      case 'getUsers':
-        return await this.getUsers(options);
-      
-      case 'createUser':
-        return await this.createUser(payload);
-      
-      case 'editUser':
-        return await this.editUser(payload);
-      
-      case 'deleteUser':
-        return await this.deleteUser(payload);
-      
-      case 'getCurrentUser':
-        return await this.getCurrentUser();
-      
-      case 'getUserSettings':
-        return await this.getUserSettings();
-      
-      case 'getUserSetting':
-        return await this.getUserSetting(payload);
-      
-      case 'setUserSetting':
-        return await this.setUserSetting(payload);
-      
-      // Recipe operations
-      case 'addRecipeProductsToShoppingList':
-        return await this.addRecipeProductsToShoppingList(payload);
-      
-      case 'getRecipeFulfillment':
-        return await this.getRecipeFulfillment(payload);
-      
-      case 'consumeRecipe':
-        return await this.consumeRecipe(payload);
-      
-      case 'getAllRecipesFulfillment':
-        return await this.getAllRecipesFulfillment(options);
-      
-      // Calendar operations
-      case 'getCalendar':
-        return await this.getCalendar();
-      
-      case 'getCalendarSharingLink':
-        return await this.getCalendarSharingLink();
-      
-      default:
-        throw new Error(`Unknown generic operation: ${operation}`);
+    // Get the appropriate module for this operation
+    const operationModule = this.operationModuleMap.get(operation);
+    if (!operationModule) {
+      throw new Error(`Unknown generic operation: ${operation}`);
     }
-  }
 
-  // System operations
-  async getSystemInfo() {
-    return await this.api.getSystemInfo();
-  }
-
-  async getDbChangedTime() {
-    return await this.api.getDbChangedTime();
-  }
-
-  async getConfig() {
-    return await this.api.getConfig();
-  }
-
-  async getTime(payload = {}) {
-    return await this.api.getTime(payload.offset);
-  }
-
-  // Generic entity operations
-  async getObjects(payload, options = {}, entityType) {
-    const entity = payload.entity || entityType;
-    Validators.validateEntity(entity);
-    const validatedOptions = Validators.validateOptions(options);
+    // Execute operation using the appropriate module
+    // For entity operations, pass the entityType parameter
+    if (this.entities.getSupportedOperations().includes(operation)) {
+      return await operationModule[operation](payload, entityType);
+    } else if (this.files.getSupportedOperations().includes(operation)) {
+      return await operationModule[operation](payload);
+    } else if (this.system.getSupportedOperations().includes(operation)) {
+      return await operationModule[operation](payload);
+    } else if (this.settings.getSupportedOperations().includes(operation)) {
+      return await operationModule[operation](payload);
+    }
     
-    return await this.api.getObjects(entity, validatedOptions);
-  }
-
-  async addObject(payload, entityType) {
-    const entity = payload.entity || entityType;
-    Validators.validateEntity(entity);
-    Validators.validateRequired(payload, ['data']);
-    Validators.validateData(payload.data);
-    
-    return await this.api.addObject(entity, payload.data);
-  }
-
-  async getObject(payload, entityType) {
-    const entity = payload.entity || entityType;
-    Validators.validateEntity(entity);
-    Validators.validateRequired(payload, ['objectId']);
-    Validators.validateId(payload.objectId, 'objectId');
-    
-    return await this.api.getObject(entity, payload.objectId);
-  }
-
-  async editObject(payload, entityType) {
-    const entity = payload.entity || entityType;
-    Validators.validateEntity(entity);
-    Validators.validateRequired(payload, ['objectId', 'data']);
-    Validators.validateId(payload.objectId, 'objectId');
-    Validators.validateData(payload.data);
-    
-    return await this.api.editObject(entity, payload.objectId, payload.data);
-  }
-
-  async deleteObject(payload, entityType) {
-    const entity = payload.entity || entityType;
-    Validators.validateEntity(entity);
-    Validators.validateRequired(payload, ['objectId']);
-    Validators.validateId(payload.objectId, 'objectId');
-    
-    return await this.api.deleteObject(entity, payload.objectId);
-  }
-
-  // User fields operations
-  async getUserfields(payload) {
-    Validators.validateRequired(payload, ['entity', 'objectId']);
-    Validators.validateEntity(payload.entity);
-    Validators.validateId(payload.objectId, 'objectId');
-    
-    return await this.api.getUserfields(payload.entity, payload.objectId);
-  }
-
-  async setUserfields(payload) {
-    Validators.validateRequired(payload, ['entity', 'objectId', 'data']);
-    Validators.validateEntity(payload.entity);
-    Validators.validateId(payload.objectId, 'objectId');
-    Validators.validateData(payload.data);
-    
-    return await this.api.setUserfields(payload.entity, payload.objectId, payload.data);
-  }
-
-  // File operations
-  async getFile(payload) {
-    Validators.validateFileParams(payload);
-    const options = payload.options || {};
-    
-    return await this.api.getFile(payload.group, payload.fileName, options);
-  }
-
-  async uploadFile(payload) {
-    Validators.validateFileParams(payload);
-    Validators.validateRequired(payload, ['fileData']);
-    
-    return await this.api.uploadFile(payload.group, payload.fileName, payload.fileData);
-  }
-
-  async deleteFile(payload) {
-    Validators.validateFileParams(payload);
-    
-    return await this.api.deleteFile(payload.group, payload.fileName);
-  }
-
-  // User operations
-  async getUsers(options = {}) {
-    const validatedOptions = Validators.validateOptions(options);
-    return await this.api.getUsers(validatedOptions);
-  }
-
-  async createUser(payload) {
-    Validators.validateData(payload);
-    return await this.api.createUser(payload);
-  }
-
-  async editUser(payload) {
-    Validators.validateRequired(payload, ['userId', 'data']);
-    Validators.validateId(payload.userId, 'userId');
-    Validators.validateData(payload.data);
-    
-    return await this.api.editUser(payload.userId, payload.data);
-  }
-
-  async deleteUser(payload) {
-    Validators.validateRequired(payload, ['userId']);
-    Validators.validateId(payload.userId, 'userId');
-    
-    return await this.api.deleteUser(payload.userId);
-  }
-
-  async getCurrentUser() {
-    return await this.api.getCurrentUser();
-  }
-
-  async getUserSettings() {
-    return await this.api.getUserSettings();
-  }
-
-  async getUserSetting(payload) {
-    Validators.validateRequired(payload, ['settingKey']);
-    Validators.validateSettingKey(payload.settingKey);
-    
-    return await this.api.getUserSetting(payload.settingKey);
-  }
-
-  async setUserSetting(payload) {
-    Validators.validateRequired(payload, ['settingKey', 'data']);
-    Validators.validateSettingKey(payload.settingKey);
-    Validators.validateData(payload.data);
-    
-    return await this.api.setUserSetting(payload.settingKey, payload.data);
-  }
-
-  // Recipe operations
-  async addRecipeProductsToShoppingList(payload) {
-    Validators.validateRequired(payload, ['recipeId', 'data']);
-    Validators.validateId(payload.recipeId, 'recipeId');
-    Validators.validateData(payload.data);
-    
-    return await this.api.addRecipeProductsToShoppingList(payload.recipeId, payload.data);
-  }
-
-  async getRecipeFulfillment(payload) {
-    Validators.validateRequired(payload, ['recipeId']);
-    Validators.validateId(payload.recipeId, 'recipeId');
-    
-    return await this.api.getRecipeFulfillment(payload.recipeId);
-  }
-
-  async consumeRecipe(payload) {
-    Validators.validateRequired(payload, ['recipeId']);
-    Validators.validateId(payload.recipeId, 'recipeId');
-    
-    return await this.api.consumeRecipe(payload.recipeId);
-  }
-
-  async getAllRecipesFulfillment(options = {}) {
-    const validatedOptions = Validators.validateOptions(options);
-    return await this.api.getAllRecipesFulfillment(validatedOptions);
-  }
-
-  // Calendar operations
-  async getCalendar() {
-    return await this.api.getCalendar();
-  }
-
-  async getCalendarSharingLink() {
-    return await this.api.getCalendarSharingLink();
+    // Fallback - shouldn't reach here
+    throw new Error(`Unknown generic operation: ${operation}`);
   }
 
   /**
@@ -307,36 +87,15 @@ class GenericOperations {
    * @returns {Array<string>} List of supported operations
    */
   getSupportedOperations() {
-    return [
-      'getSystemInfo',
-      'getDbChangedTime',
-      'getConfig',
-      'getTime',
-      'getObjects',
-      'addObject',
-      'getObject',
-      'editObject',
-      'deleteObject',
-      'getUserfields',
-      'setUserfields',
-      'getFile',
-      'uploadFile',
-      'deleteFile',
-      'getUsers',
-      'createUser',
-      'editUser',
-      'deleteUser',
-      'getCurrentUser',
-      'getUserSettings',
-      'getUserSetting',
-      'setUserSetting',
-      'addRecipeProductsToShoppingList',
-      'getRecipeFulfillment',
-      'consumeRecipe',
-      'getAllRecipesFulfillment',
-      'getCalendar',
-      'getCalendarSharingLink'
-    ];
+    // Collect operations from all specialized modules
+    const allOperations = [];
+    
+    allOperations.push(...this.entities.getSupportedOperations());
+    allOperations.push(...this.files.getSupportedOperations());
+    allOperations.push(...this.system.getSupportedOperations());
+    allOperations.push(...this.settings.getSupportedOperations());
+    
+    return allOperations;
   }
 }
